pyproject.toml
[project]
name = "git-local"
version = "0.1.0"
description = "Local GitLab-style overview for Git repositories"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "gitpython>=3.1.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
git-local = "git_local.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
